+++ app/build.gradle.kts
@@ -57,13 +57,19 @@ dependencies {
     implementation ("com.google.android.gms:play-services-location:17.0.0")
     implementation ("com.github.bumptech.glide:glide:4.16.0")
 
-
-
     implementation(platform("com.google.firebase:firebase-bom:33.1.2"))
     implementation("com.google.firebase:firebase-analytics")
+
+
+
+
+
+
+
+
+
+    //implementaciones nuevas
     implementation("com.google.firebase:firebase-firestore-ktx")
-
-
     implementation("com.google.firebase:firebase-auth-ktx")
     implementation ("com.google.firebase:firebase-storage-ktx")
