+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
             android:value="@string/title_activity_maps" />
 
         <activity
-            android:name=".MainActivity"
+            android:name=".view.ui.MainActivity"
             android:exported="true"
             android:label="@string/title_activity_maps">
             <intent-filter>
