--- app/src/main/java/com/example/location_feature/adapter/horarioViewHolder.java
@@ -1,4 +1,0 @@
-package com.example.location_feature.adapter;
-
-public class horarioViewHolder {
-}
