+++ app/src/main/java/com/example/location_feature/domain/model/Comentario.kt
@@ -1,0 +1,38 @@
+package com.example.location_feature.domain.model
+
+
+import android.os.Parcel
+import android.os.Parcelable
+
+data class Comentario(
+    var id: String = "",
+    var tema: String = "",
+    var comentario: String = ""
+) : Parcelable {
+
+    constructor(parcel: Parcel) : this(
+        id = parcel.readString() ?: "",
+        tema = parcel.readString() ?: "",
+        comentario = parcel.readString() ?: ""
+    )
+
+    override fun writeToParcel(parcel: Parcel, flags: Int) {
+        parcel.writeString(id)
+        parcel.writeString(tema)
+        parcel.writeString(comentario)
+    }
+
+    override fun describeContents(): Int {
+        return 0
+    }
+
+    companion object CREATOR : Parcelable.Creator<Comentario> {
+        override fun createFromParcel(parcel: Parcel): Comentario {
+            return Comentario(parcel)
+        }
+
+        override fun newArray(size: Int): Array<Comentario?> {
+            return arrayOfNulls(size)
+        }
+    }
+}
