+++ app/src/main/java/com/example/location_feature/domain/model/Usuario.kt
@@ -1,0 +1,8 @@
+package com.example.location_feature.domain.model
+
+data class Usuario(
+    var id: String = "",
+    var email: String = "",
+    var password: String = "",
+    var rol: String = "",
+)
