+++ app/src/main/java/com/example/location_feature/domain/repository/AuthRepository.kt
@@ -1,0 +1,7 @@
+package com.example.location_feature.domain.repository
+
+interface AuthRepository {
+
+        suspend fun login(email: String, password:String): Boolean
+        suspend fun signUp(email:String, password: String): Boolean
+}
