+++ app/src/main/java/com/example/location_feature/domain/usercase/FirebaseLoginCase.kt
@@ -1,0 +1,23 @@
+package com.example.location_feature.domain.usercase
+
+import com.example.location_feature.domain.repository.AuthRepository
+import com.example.location_feature.util.Resource
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import javax.inject.Inject
+
+class FirebaseLoginCase @Inject constructor(
+    private val authRepository: AuthRepository
+){
+    suspend operator fun invoke(email: String, password: String): Flow<Resource<Boolean>> = flow {
+        emit(Resource.Loading)
+        val loggedSuccessfully = authRepository.login(email,password)
+        if (loggedSuccessfully) {
+            emit(Resource.Success(true))
+            emit(Resource.Finished)
+        } else {
+            emit(Resource.Error("Login error"))
+            emit(Resource.Finished)
+        }
+    }
+}
