--- app/src/main/java/com/example/location_feature/model/Conductor.kt
@@ -1,14 +1,0 @@
-package com.example.location_feature.model
-
-import android.os.Parcel
-import android.os.Parcelable
-
-data class Conductor(
-    var id: String = "",
-    var nombres: String = "",
-    var apellidos: String = "",
-    var email: String = "",
-    var clave: String = "",
-    var camionid: String = ""
-
-)
