+++ app/src/main/java/com/example/location_feature/network/FirebaseAuthRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.example.location_feature.network
 
 import android.util.Log
-import com.example.location_feature.repository.AuthRepository
+import com.example.location_feature.domain.repository.AuthRepository
 import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
@@ -10,10 +10,10 @@ class FirebaseAuthRepositoryImpl @Inject constructor(
     private val firebaseAuth: FirebaseAuth
 ): AuthRepository {
 
-    override suspend fun login(correo: String, clave: String): Boolean {
+    override suspend fun login(email: String, password: String): Boolean {
         return try {
             var isSuccessful: Boolean = false
-            firebaseAuth.signInWithEmailAndPassword(correo, clave)
+            firebaseAuth.signInWithEmailAndPassword(email, password)
                 .addOnSuccessListener { isSuccessful = true }
                 .addOnFailureListener { isSuccessful = false }
                 .await()
@@ -24,10 +24,10 @@ class FirebaseAuthRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun signUp(correo: String, clave: String): Boolean {
+    override suspend fun signUp(email: String, password: String): Boolean {
         return try {
             var isSuccessful: Boolean = false
-            firebaseAuth.createUserWithEmailAndPassword(correo, clave)
+            firebaseAuth.createUserWithEmailAndPassword(email, password)
                 .addOnCompleteListener { isSuccessful = it.isSuccessful}
                 .await()
             isSuccessful
