+++ app/src/main/java/com/example/location_feature/view/adapter/horarioAdapter.java
@@ -1,0 +1,4 @@
+package com.example.location_feature.view.adapter;
+
+public class horarioAdapter {
+}
