+++ app/src/main/java/com/example/location_feature/view/password_recovery/PasswordRecoveryFragment.kt
@@ -1,0 +1,29 @@
+package com.esaudev.firebaseyt.presentation.password_recovery
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.esaudev.firebaseyt.databinding.FragmentPasswordRecoveryBinding
+
+class PasswordRecoveryFragment : Fragment() {
+
+    private var _binding: FragmentPasswordRecoveryBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        // Inflate the layout for this fragment
+        _binding = FragmentPasswordRecoveryBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+}
