--- app/src/main/res/layout/comentarios_cuenta.xml
@@ -1,163 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/comentarios"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#ffffff"
-
-    >
-
-    <ImageView
-        android:id="@+id/image_view_comentarios"
-        android:layout_width="450dp"
-        android:layout_height="140dp"
-        android:adjustViewBounds="true"
-        android:scaleType="fitXY"
-        android:src="@drawable/shape_comentario"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.512"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageButton
-        android:id="@+id/btRegreo2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="26dp"
-        android:layout_marginTop="44dp"
-        android:background="@drawable/shape_regresar_notificacion"
-        android:contentDescription="butonback"
-        android:padding="15dp"
-        android:src="@drawable/ic_arrow_back_notificacion_comentarios"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/Tlcomentario"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/Tlcomentario"
-
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="40dp"
-        android:layout_marginTop="52dp"
-        android:text="Comentarios"
-        android:textColor="#000000"
-        android:textSize="25sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toEndOf="@+id/btRegreo2"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/subtlcomentario"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="24dp"
-        android:layout_marginTop="72dp"
-        android:text="Tu Opinión nos Importa"
-        android:textColor="#000000"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/btRegreo2" />
-
-    <TextView
-        android:id="@+id/texto_Escríbenos"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:text="Escríbenos tu comentario, así nos ayudas a mejorar."
-        android:textSize="12sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/subtlcomentario" />
-
-    <!-- TextView -->
-    <TextView
-        android:id="@+id/texto_tema"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="25dp"
-        android:layout_marginTop="16dp"
-        android:textStyle="bold"
-        android:text="Tema"
-        android:textSize="16sp"
-        android:textColor="#000000"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/texto_Escríbenos" />
-
-    <!-- EditText -->
-    <EditText
-        android:id="@+id/tv_texto_commentario"
-        android:layout_width="350dp"
-        android:layout_height="65dp"
-        android:layout_marginStart="25dp"
-        android:layout_marginTop="16dp"
-        android:background="@drawable/shape_cudro_texto_comentario"
-        android:gravity="top"
-        android:hint="Escribe tu tema"
-        android:inputType="textMultiLine"
-        android:lines="4"
-        android:minHeight="80dp"
-        android:paddingStart="12dp"
-        android:paddingTop="8dp"
-        android:paddingEnd="12dp"
-        android:paddingBottom="8dp"
-        android:textColor="#000000"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/texto_tema" />
-
-    <!-- TextView -->
-    <TextView
-        android:id="@+id/texto_comentario"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="25dp"
-        android:layout_marginTop="16dp"
-        android:textStyle="bold"
-        android:text="Escribe tu comentario"
-        android:textSize="16sp"
-        android:textColor="#000000"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_texto_commentario" />
-
-    <!-- EditText -->
-    <EditText
-        android:id="@+id/tv_texto_commentario2"
-        android:layout_width="350dp"
-        android:layout_height="150dp"
-        android:layout_marginStart="25dp"
-        android:layout_marginTop="16dp"
-        android:background="@drawable/shape_cudro_texto_comentario"
-        android:gravity="top"
-        android:hint="Escribe tu comentario"
-        android:inputType="textMultiLine"
-        android:lines="4"
-        android:minHeight="80dp"
-        android:paddingStart="12dp"
-        android:paddingTop="8dp"
-        android:paddingEnd="12dp"
-        android:paddingBottom="8dp"
-        android:textColor="#000000"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/texto_comentario" />
-
-    <!-- Button -->
-    <Button
-        android:id="@+id/btnEnvia_comentarior"
-        android:layout_width="160dp"
-        android:layout_height="80dp"
-        android:background="#14FF85"
-        android:text="ENVIAR"
-        android:textColor="#000000"
-        android:textSize="25sp"
-        android:textStyle="bold"
-        android:gravity="center"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_texto_commentario2"
-        android:layout_marginStart="115dp"
-        android:layout_marginTop="30dp" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
