+++ app/src/main/res/layout/fragment_home_mapa.xml
@@ -6,4 +6,4 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vistas.MapsActivity" />
+    tools:context=".view.ui.Activities.MapsActivity" />
