+++ app/src/main/res/layout/fragment_notifecaciones.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity"
+    tools:context=".view.ui.MainActivity"
     android:background="#ffffff"
     android:id="@+id/notificacionbot"
     xmlns:app="http://schemas.android.com/apk/res-auto">
